% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{nav_content}
\alias{nav_content}
\title{Nav content}
\usage{
nav_content(
  ...,
  ids = NULL,
  container_class = NULL,
  content_class = NULL,
  tabsetid = "tabSet1"
)
}
\arguments{
\item{...}{to be added}

\item{ids}{to be added}

\item{container_class}{to be added}

\item{content_class}{to be added}

\item{tabsetid}{to be added}
}
\description{
Credit: DavZim https://github.com/rstudio/bslib/issues/585
}
