% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{nav_tab}
\alias{nav_tab}
\title{Nav tab}
\usage{
nav_tab(..., ids = NULL, tabsetid = "tabSet1")
}
\arguments{
\item{...}{to be added}

\item{ids}{to be added}

\item{tabsetid}{to be added}
}
\description{
Credit: DavZim https://github.com/rstudio/bslib/issues/585
}
