% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\docType{data}
\name{nav_tab_js}
\alias{nav_tab_js}
\title{Java Script for nav_tab function}
\format{
An object of class \code{character} of length 1.
}
\usage{
nav_tab_js
}
\description{
Credit: DavZim https://github.com/rstudio/bslib/issues/585
}
\keyword{datasets}
