% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{remove_field}
\alias{remove_field}
\title{Remove field if in data frame if it exists}
\usage{
remove_field(df, x)
}
\arguments{
\item{df}{data frame}

\item{x}{field to be removed if exists.}
}
\description{
Remove field if in data frame if it exists
}
