% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytemplate.R
\name{shinytemplateDev}
\alias{shinytemplateDev}
\title{shinytemplate runs the app in development mode}
\usage{
shinytemplateDev(option1 = NULL, option2 = NULL)
}
\arguments{
\item{option1}{Option 1}

\item{option2}{Option 2}
}
\description{
shinytemplate runs the app in development mode
}
