% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytemplate.R
\name{shinytemplate}
\alias{shinytemplate}
\title{Launch the shinytemplate app}
\usage{
shinytemplate(option1 = NULL, option2, dev = F)
}
\arguments{
\item{option1}{An option for your app}

\item{option2}{Another option for your app}

\item{dev}{Loads from local directories if selected}
}
\description{
Executing this function will launch the shinytemplate application in
the user's default web browser.
}
\examples{
\dontrun{
shinytemplate(option1 = 'hello', option2 = 'world')
}
}
