% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytemplate.R
\name{shinytemplate}
\alias{shinytemplate}
\title{Launch the shinytemplate app}
\usage{
shinytemplate(option1 = NULL, option2 = NULL)
}
\arguments{
\item{option1}{Option 1}

\item{option2}{Option 2}
}
\description{
Executing this function will launch the shinytemplate application in
the user's default web browser.
}
\examples{
\dontrun{
shinytemplate()
}
shinytemplate runs the app in production mode
}
\author{
Colin Diesh \email{colin.diesh@gmail.com}
}
